Added: Drag and Drop
Might be slightly crude at the moment but this enables the user to move
equiment around.

To do: Persist the locations - WIll need to enable a mongodb database
for this and set up routes in node, save and retrieve the objects.

Bugs: The mouse position goes off when clicking an object. the y axis is
broken when trying to change the position. The scroll bar messes it up
but fixes again when scrolling back