Updated: Dashboard -> StoreEquipment
Dashboard moved to store equipment. This is where the map of the stock
room opens up. This component loads the equipment attached to the
specific store.

To do: So it loads up all the pieces of equipment, but it doesnt
dynamically add the styles to it. Need to figure out how to add styles
that are saved on an object.

Possible solutions;
Loop through each piece of equipment and use the Renderer to create each
element and then add styles to that element.

Split the making of the elements into a component and dynamically load
the component. Need tp figure out how to pass the object data between
components.